backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
django-tastypie==0.14.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pep8==1.7.1
postgres==3.0.0
psycopg2==2.8.4
psycopg2-pool==1.1
pylint==2.4.4
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
django-cors-headers==3.2.1  
